backend/requirements.txt
flask>=2.0.1
flask-cors>=4.0.0
Werkzeug>=2.0.3
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
python-dotenv>=1.0.0
pathlib>=1.0.1
ultralytics>=8.0.196
